Fallback to local composer (WIP)
When working offline, the build.sh script failed, because it's not
possible to obtain composer.phar from the webserver.

Now the build falls back to the local copy of "composer".

This is work in progress because while the detection for the local composer
works, the build script itself is not yet flexible with it as it relates
to composer.phar as file.

TODO: Locate local composer.phar from the environment and then copy it
      to the build location (instead of changing the location in the
      script)

Conflicts:
	build.sh